Move idc viewer to proteinpaint folder.
